refactor: decompose search component into modular atomic sub-components
