+++ src/components/layout/Header.astro
@@ -2,7 +2,7 @@
 import { Image } from "astro:assets";
 import logoDark from "@/assets/images/logo-dark.svg";
 import logoLight from "@/assets/images/logo-light.svg";
-import Search from "@/components/ui/Search.svelte";
+import Search from "@/components/ui/search/Search.svelte";
 import ThemeToggle from "@/components/ui/ThemeToggle.astro";
 import siteConfig from "@/config/site-config";
